refactor(runtime): split runtime responsibilities
