+++ packages/runtime/src/index.ts
@@ -1,8 +1,8 @@
-export * from "./domain/types.js";
+export * from "./domain/index.js";
 export * from "./runtime.js";
-export * from "./diff/diff.js";
+export * from "./timeline/snapshot-diff.js";
 export * from "./verification/verification.js";
 export * from "./editing/intent.js";
 export * from "./capabilities.js";
-export * from "./snapshot-digest.js";
+export * from "./timeline/snapshot-digest.js";
 export * from "./speech/filler-removal.js";
